Create Role Changer
For convenience, this function should be detect the meaning of a string and change a player's role to the given role. It kinda seems bad to do it this way, but it has its advantages; this way, one can exactly control what happens. For example, if a player turns into the Pyromancer, their powdered tag can immediately be removed, preventing them from accidentally getting killed. Also, the Exorcist cannot be demonized, as it is always set to False whenever one turns into the Exorcist.