Added Message class
Instead of having to record all the changes and send them back, I thought it'd be neat if all changes were made immediately, and that only the messages that need to be sent are being returned. This is why all functions should return Mailbox classes instead of True/False booleans or roles.