Added tanner's "fakerole"
When a crowd seeker, fortune teller or warlock is inspecting a participant, that is the role that should be returned, not the actual role.
For future reference, a check WILL be needed if the role and fakerole match and undead is set to True. In that case, it should display as undead.