Added channel table
To prevent time, the bot should save some channels in history that it had used recently to send messages. This way, the bot will only need to ask channel objects once from the API.
Naturally, after doing a hard reboot of the bot, the table will be forgotten and refilled.