Created a dynamic config file
This file should contain information that isn't static, like the information and settings given in config.py, but is too small to create a whole new table for in the database. These are small pieces (about the size of O(1) of data) of information that will be accessed by the bot and manually managed by the Game Masters, if necessary.