Final test fix
For some reason, mr. Evil One understood that it wasn't wise not to cooperate, so it seems to work again! It did however point out a little thing that I do need to update whenever I update the database. It's not a part of the bot, so it's not a problem, but it's still something that needs to be regarded anyway.