Add functionality for secret channel pointers
This is the most efficient way to implement secret channels without copypasting the whole cc system. To implement secret channels into the cc system, the secret channel's owner is set as an imaginary user, and a separate table in the SQLite database called `secret_channel` keeps track of the channels and to whom they belong.
The only thing that then needs to be done separately, is the table that keeps track of categories. This isn't necessary, though it looks much better if the cc channels don't mix up with the secret channels. That will be something that yet needs to be fixed.