Set Amulet Holder's channel
The system to remember amulets is no longer storing who has what amulet, but instead keeps track of the channels by the Amulet Holder's role. This way, the Amulet Holder can change roles and still get the amulet back, AND the bot doesn't risk gaining endless lists of amulets that need to be parsed to gain the requested information.