Add more death mechanics to role_data.py
The Cursed Civilian will now turn into a Werewolf when attacked by a wolf. 
The Horsemen can now apocalypse non-horsemen. 
Players can no longer die if they are frozen or abducted. 
Demonized players can now turn undead. *(except for a few exceptions)*

The suicide command, and this is the most important part, needs to be desgined in such a way that it is easy to duplicate, so users can easily take others with them.
I think I'll combine the functions written in this file to make the functions recursive. The levels shouldn't go too deep, meaning this is an elegant and easy way to solve kills.